field-manager-python-client/pyproject.toml
[tool.poetry]
name = "field-manager-python-client"
version = "4.6.89"
description = "A client library for accessing Field Manager Data API"
authors = ["Jiyang Ye <jiyang.ye@ngi.no>"]
maintainers = ["Jiyang Ye <jiyang.ye@ngi.no>", "Jostein Leira <jostein.leira@ngi.no>"]
readme = "README.md"
packages = [
    {include = "field_manager_python_client"},
]
include = ["CHANGELOG.md", "field_manager_python_client/py.typed"]


[tool.poetry.dependencies]
python = "^3.9"
httpx = ">=0.23.0,<0.29.0"
attrs = ">=22.2.0"
python-dateutil = "^2.8.0"
python-keycloak = "^5.5.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["F", "I", "UP"] 